requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
crispy-bootstrap5==0.7
cryptography==41.0.4
deep-translator==1.11.4
defusedxml==0.8.0rc2
Django==4.2.5
django-crispy-forms==2.0
django-extensions==3.2.3
django-ranged-response==0.2.0
django-simple-captcha==0.5.20
django-taggit==4.0.0
idna==3.4
oauthlib==3.2.2
Pillow==10.0.1
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.3.0
social-auth-core==4.4.2
soupsieve==2.5
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.6
